backend/user_service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
pydantic-extra-types==2.0.0
email-validator==2.0.0.post2
python-jose>=3.3.0
python-multipart>=0.0.6
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy>=2.0.9
alembic>=1.10.3
psycopg2-binary>=2.9.6
pyyaml==6.0.1
jsonschema==4.18.6
pyjwt>=2.6.0
aiohttp==3.8.5
pillow==10.0.0
python-magic==0.4.27
tenacity==8.2.2
python-dotenv>=1.0.0
asyncpg>=0.27.0
httpx>=0.24.0
